+++ CHANGELOG.md
@@ -22,6 +22,7 @@ New features:
 Other improvements:
 - Docs: updated package addition/overriding syntax to use `with` syntax (#661)
 - Pass main function argument to `--run` for alternate backends
+- Support `upgrade set` for alternative package-set repositories
 
 ## [0.15.3] - 2020-06-15